Add fetchApi file
