+++ apps/charterafrica/payload.config.js
@@ -7,7 +7,6 @@ import seo from "@payloadcms/plugin-seo";
 import dotenv from "dotenv";
 import { buildConfig } from "payload/config";
 
-import Errors from "./src/payload/collections/Errors";
 import Explainers from "./src/payload/collections/Explainers";
 import Media from "./src/payload/collections/Media";
 import Pages from "./src/payload/collections/Pages";
@@ -53,7 +52,7 @@ const adapter = s3Adapter({
 
 export default buildConfig({
   serverURL: appURL,
-  collections: [Errors, Explainers, Media, Pages, Partners, Resources],
+  collections: [Explainers, Media, Pages, Partners, Resources],
   globals: [FocalCountries, Footer, Helpdesk, Navigation, Settings],
   ...(locales?.length
     ? {
